refactor: extract helpers.py
